Provide option to specify CRN with Python code
If the SOLVERS_ENABLED environmental variable is true, then the
project form will provide a textarea into which a block of python
code can be pasted. When the form is submitted, this will be run
with exec() in order to create the CRN object.

This has obvious security implications, and should only be enabled
when registration of new users is disabled and all existing users
are trusted.

Long term, the Right Thing to do is to instead parse textual
descriptions of CRN sketches, but there isn't time to implement
this now.